[Driver] PriceFinder that uses liner optimization solver (#75)
This PR adds an implementation of the price finding interface that calls into the linear optimization solver @twalth3r implemented in Python.

The component is doing the following steps:

Convert Orders/Balances into the json format the solver expects
1. Write json to disk
2. Invoke solver passing the json file that was written
3. Reading the result file into json
4. Converting the result file into solution struct (and storing prices in memory for next run)

I added a custom error class for PriceFinding, which should help debugging in the future (since we will be able to distinguish other error from solver related errors). Once used, that error should probably be wrapped into a DriverError.

I'm planning to add unit tests for the serialization/deserialization methods, but would suggest doing that in the following PR, since this one is already quite big.

To test for now, I added the following code inside the main method:

```rust
    let price_finder = LinearOptimisationPriceFinder::new();
    let orders = vec![Order {
        account_id: 0,
        sell_token: 0,
        buy_token: 1,
        sell_amount: 1000000000000000000,
        buy_amount: 1000000000000000000
    }, Order {
        account_id: 0,
        sell_token: 1,
        buy_token: 0,
        sell_amount: 1000000000000000000,
        buy_amount: 1000000000000000000
    }];
    let state = State {
        state_hash: "".to_string(),
        state_index: 0,
        balances: vec![1000000000000000000; 60],
    };
    price_finder.find_prices(&orders, &state).expect("Price Finding Failed");
```

And see that the result is computed correctly

+++ driver/src/deposit_driver.rs
@@ -12,7 +12,7 @@ pub fn apply_deposits(
     deposits: &Vec<models::PendingFlux>,
 ) -> models::State {
     for i in deposits {
-        state.balances[((i.account_id - 1) * models::TOKENS + (i.token_id as u16 - 1)) as usize] += i.amount;
+        state.balances[((i.account_id - 1) * (models::TOKENS as u16) + (i.token_id as u16 - 1)) as usize] += i.amount;
     }
     state.clone()
 }
